Refactoring

+++ src/transform/segment.rs
@@ -4,13 +4,13 @@ use crate::Vector;
 #[derive(derive_more::Constructor)]
 pub struct SegmentsFactory {
     child: Box<dyn PatternFactory>,
-    segments: u8,
+    segments: f32,
 }
 
 #[derive(derive_more::Constructor)]
 pub struct Segments {
     child: Box<dyn Pattern>,
-    segments: u8,
+    segments: f32,
 }
 
 impl PatternFactory for SegmentsFactory {
@@ -24,7 +24,7 @@ impl Pattern for Segments {
         let sample = self.child.sample(pos);
 
         if 0.0 <= sample && sample < 1.0 {
-            (sample * self.segments as f32) % 1.0
+            sample * self.segments % 1.0
         } else {
             sample
         }
@@ -51,7 +51,7 @@ mod test {
             .once()
             .returning(|_| Box::new(MockPattern::new()));
 
-        SegmentsFactory::new(Box::new(child), 2).create(&config);
+        SegmentsFactory::new(Box::new(child), 2.0).create(&config);
     }
 
     #[test]
@@ -63,7 +63,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 3).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 3.0).create(&Config::default());
 
         assert_abs_diff_eq!(1.1, sampler.sample(Vector::default()));
     }
@@ -77,7 +77,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 3).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 3.0).create(&Config::default());
 
         assert_abs_diff_eq!(-0.1, sampler.sample(Vector::default()));
     }
@@ -91,7 +91,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 4).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 4.0).create(&Config::default());
 
         assert_abs_diff_eq!(0.96, sampler.sample(Vector::default()), epsilon = 0.01);
     }
@@ -105,7 +105,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 4).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 4.0).create(&Config::default());
 
         assert_abs_diff_eq!(0.0, sampler.sample(Vector::default()));
     }
@@ -119,7 +119,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 4).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 4.0).create(&Config::default());
 
         assert_eq!(0.96, sampler.sample(Vector::default()));
     }
@@ -133,7 +133,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 4).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 4.0).create(&Config::default());
 
         assert_abs_diff_eq!(0.0, sampler.sample(Vector::default()));
     }
@@ -151,7 +151,7 @@ mod test {
             Box::new(sampler)
         });
 
-        let sampler = SegmentsFactory::new(Box::new(child), 3).create(&Config::default());
+        let sampler = SegmentsFactory::new(Box::new(child), 3.0).create(&Config::default());
 
         sampler.sample(Vector::new(5.0, 1.0));
     }
